+++ tests/readme.rs
@@ -1,18 +1,5 @@
-# `impl-more`
+#![allow(dead_code, clippy::from_over_into)]
 
-> Concise trait implementations.
-
-[![crates.io](https://img.shields.io/crates/v/impl-more?label=latest)](https://crates.io/crates/impl-more)
-[![Documentation](https://docs.rs/impl-more/badge.svg)](https://docs.rs/impl-more/0.1.4)
-![MIT or Apache 2.0 licensed](https://img.shields.io/crates/l/impl-more.svg)
-<br />
-[![dependency status](https://deps.rs/crate/impl-more/0.1.4/status.svg)](https://deps.rs/crate/impl-more/0.1.4)
-[![Download](https://img.shields.io/crates/d/impl-more.svg)](https://crates.io/crates/impl-more)
-[![CircleCI](https://circleci.com/gh/robjtede/impl-more/tree/main.svg?style=shield)](https://circleci.com/gh/robjtede/impl-more/tree/main)
-
-# Usage
-
-```rust
 struct MyNewTypeStruct(String);
 
 impl_more::impl_as_ref!(MyNewTypeStruct => String);
@@ -52,4 +39,3 @@ enum Err {
 
 impl_more::impl_display_enum!(Err, Io(err) => "{err}", Generic(msg) => "{msg}");
 impl_more::impl_error_enum!(Err, Io(err) => err);
-```